requirements.txt
fastapi>=0.109.0
python-multipart>=0.0.6
uvicorn[standard]>=0.27.0
google-auth>=2.27.0
google-api-python-client>=2.115.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.26.0
aiogram>=3.13.0
jinja2>=3.1.0
itsdangerous>=2.1.0
